added colour support for semesters
