+++ packages/frontend/app/(dashboard)/organization/settings/page.tsx
@@ -8,14 +8,11 @@ import {
   Form,
   Input,
   message,
-  Modal,
   Row,
-  Select,
   Spin,
   Switch,
 } from 'antd'
 import TextArea from 'antd/es/input/TextArea'
-import dayjs from 'dayjs'
 import { ReactElement, useEffect, useState } from 'react'
 import { useUserInfo } from '@/app/contexts/userContext'
 import { Organization } from '@/app/typings/organization'
@@ -23,8 +20,7 @@ import { API } from '@/app/api'
 import Image from 'next/image'
 import ImageCropperModal from '@/app/(dashboard)/components/ImageCropperModal'
 import { OrganizationRole, SemesterPartial } from '@koh/common'
-import { SemesterModal } from './components/SemesterModal'
-import DeleteConfirmationModal from './components/DeleteConfirmationModal'
+import { SemesterManagement } from './components/SemesterManagement'
 
 export default function SettingsPage(): ReactElement {
   const [formGeneral] = Form.useForm()
@@ -49,15 +45,7 @@ export default function SettingsPage(): ReactElement {
   const [organizationSemesters, setOrganizationSemesters] = useState<
     SemesterPartial[]
   >([])
-  const [currentSemesterId, setCurrentSemesterId] = useState<number>(-1) // -1 represents nothing being selected
-  const [deletionSemesterName, setDeletionSemesterName] = useState<string>('')
-  const [isSemesterCreationModalOpen, setIsSemesterCreationModalOpen] =
-    useState(false)
-  const [isSemesterEditModalOpen, setIsSemesterEditModalOpen] = useState(false)
-  const [
-    isConfirmSemesterDeleteModalOpen,
-    setIsConfirmSemesterDeleteModalOpen,
-  ] = useState(false)
+
   useEffect(() => {
     const fetchDataAsync = async () => {
       const response = await API.organizations.get(
@@ -76,6 +64,7 @@ export default function SettingsPage(): ReactElement {
             startDate: new Date(s.startDate),
             endDate: new Date(s.endDate),
             description: s.description,
+            color: s.color,
           }
         }),
       )
@@ -149,152 +138,6 @@ export default function SettingsPage(): ReactElement {
         const errorMessage = error.response.data.message
 
         message.error(errorMessage)
-      })
-  }
-
-  // for semester management
-  const [semesterForm] = Form.useForm()
-
-  const handleAddSemester = async () => {
-    const formValues = await semesterForm.validateFields([
-      'name',
-      'startDate',
-      'endDate',
-    ])
-    const semesterName = formValues.name as string
-    const semesterStartDate = formValues.startDate as dayjs.Dayjs
-    const semesterEndDate = formValues.endDate as dayjs.Dayjs
-    const semesterDescription = semesterForm.getFieldValue(
-      'description',
-    ) as string
-
-    if (semesterName.length < 3) {
-      message.error('Semester name must be at least 3 characters')
-      return
-    }
-
-    if (semesterDescription && semesterDescription.length < 10) {
-      message.error('Semester description must be at least 10 characters')
-      return
-    }
-
-    if (semesterStartDate.valueOf() >= semesterEndDate.valueOf()) {
-      message.error('Semester start date must be before end date')
-      return
-    }
-
-    const semesterDetails: SemesterPartial = {
-      name: semesterName,
-      startDate: semesterStartDate.toDate(),
-      endDate: semesterEndDate.toDate(),
-      description: semesterDescription,
-    }
-
-    await API.semesters
-      .create(organization?.id ?? -1, semesterDetails)
-      .then(() => {
-        setIsSemesterCreationModalOpen(false)
-        message.success('Semester created successfully')
-        setOrganizationSemesters((prev) => [...prev, semesterDetails])
-        semesterForm.resetFields()
-      })
-      .catch((error) => {
-        message.error(error.response.data.message)
-      })
-  }
-
-  const handleOpenEditSemesterModal = (semesterId: number) => {
-    {
-      const semester = organizationSemesters.find((s) => s.id === semesterId)
-      if (!semester) {
-        message.error('Semester not found')
-        return
-      }
-      semesterForm.setFieldsValue({
-        name: semester.name,
-        startDate: dayjs(semester.startDate),
-        endDate: dayjs(semester.endDate),
-        description: semester.description,
-      })
-      setCurrentSemesterId(semesterId)
-      setIsSemesterEditModalOpen(true)
-    }
-  }
-
-  const handleEditSemester = async () => {
-    const formValues = await semesterForm.validateFields([
-      'name',
-      'startDate',
-      'endDate',
-    ])
-    const semesterName = formValues.name
-    const semesterStartDate = formValues.startDate
-    const semesterEndDate = formValues.endDate
-    const semesterDescription = semesterForm.getFieldValue('description')
-
-    if (semesterName.length < 3) {
-      message.error('Semester name must be at least 3 characters')
-      return
-    }
-
-    if (semesterDescription && semesterDescription.length < 10) {
-      message.error('Semester description must be at least 10 characters')
-      return
-    }
-
-    if (semesterStartDate >= semesterEndDate) {
-      message.error('Semester start date must be before end date')
-      return
-    }
-
-    const semesterDetails: SemesterPartial = {
-      name: semesterName,
-      startDate: semesterStartDate,
-      endDate: semesterEndDate,
-      description: semesterDescription || null,
-    }
-
-    await API.semesters
-      .edit(organization?.id ?? -1, currentSemesterId, semesterDetails)
-      .then(() => {
-        setIsSemesterEditModalOpen(false)
-        setCurrentSemesterId(-1)
-        message.success('Semester updated successfully')
-        setOrganizationSemesters((prev) => {
-          const index = prev.findIndex((s) => s.id === semesterDetails.id)
-          prev[index] = semesterDetails
-          return prev
-        })
-      })
-      .catch((error) => {
-        const errorMessage = error.response.data.message
-
-        message.error(errorMessage)
-      })
-  }
-
-  const handleConfirmSemesterDelete = (
-    semesterId: number,
-    semesterName: string,
-  ) => {
-    setCurrentSemesterId(semesterId)
-    setDeletionSemesterName(semesterName)
-    setIsConfirmSemesterDeleteModalOpen(true)
-  }
-
-  const handleDeleteSemester = async (semesterId: number) => {
-    await API.semesters
-      .delete(organization?.id ?? -1, semesterId)
-      .then(() => {
-        setCurrentSemesterId(-1)
-        setDeletionSemesterName('')
-        message.success('Semester deleted successfully')
-        setOrganizationSemesters((prev) =>
-          prev.filter((s) => s.id !== semesterId),
-        )
-      })
-      .catch((error) => {
-        message.error(error.response.data.message)
       })
   }
 
@@ -509,89 +352,11 @@ export default function SettingsPage(): ReactElement {
         </>
       )}
 
-      <Card title="Semester Management" bordered className="mb-10 w-full">
-        {organizationSemesters && organizationSemesters.length > 0 ? (
-          organizationSemesters
-            .sort((a, b) => b.startDate.getTime() - a.startDate.getTime())
-            .map((semester) => (
-              <Card.Grid
-                key={semester.id}
-                className="flex w-[50%] flex-col justify-between gap-2 text-center hover:cursor-pointer"
-                onClick={() => handleOpenEditSemesterModal(semester.id!)}
-              >
-                <h3 className="text-lg font-semibold">{semester.name}</h3>
-                <p>
-                  <span className="font-semibold">Start Date:</span>{' '}
-                  {semester.startDate.toDateString()}
-                </p>
-                <p>
-                  <span className="font-semibold">End Date:</span>{' '}
-                  {semester.endDate.toDateString()}
-                </p>
-                {semester.description && (
-                  <p>
-                    <span className="font-semibold">Description:</span>{' '}
-                    {semester.description}
-                  </p>
-                )}
-                <Button
-                  danger
-                  type="primary"
-                  onClick={(e) => {
-                    e.stopPropagation()
-                    handleConfirmSemesterDelete(semester.id!, semester.name)
-                  }}
-                  className="mt-2"
-                >
-                  Delete
-                </Button>
-              </Card.Grid>
-            ))
-        ) : (
-          <Card.Grid className="w-[50%] text-center">
-            No semesters added yet. Click the button to add a new semester.
-          </Card.Grid>
-        )}
-
-        <Card.Grid className="w-[50%] text-center">
-          <Button
-            type="dashed"
-            onClick={() => setIsSemesterCreationModalOpen(true)}
-          >
-            + Add New Semester
-          </Button>
-        </Card.Grid>
-
-        {isSemesterCreationModalOpen && (
-          <SemesterModal
-            isSemesterModalOpen={isSemesterCreationModalOpen}
-            setIsSemesterModalOpen={setIsSemesterCreationModalOpen}
-            handleSubmit={handleAddSemester}
-            semesterForm={semesterForm}
-            creatingSemester={true}
-          />
-        )}
-        {isSemesterEditModalOpen && (
-          <SemesterModal
-            isSemesterModalOpen={isSemesterEditModalOpen}
-            setIsSemesterModalOpen={setIsSemesterEditModalOpen}
-            handleSubmit={handleEditSemester}
-            semesterForm={semesterForm}
-            creatingSemester={false}
-          />
-        )}
-        {isConfirmSemesterDeleteModalOpen && (
-          <DeleteConfirmationModal
-            isOpen={isConfirmSemesterDeleteModalOpen}
-            semesterName={deletionSemesterName}
-            onConfirm={() => {
-              handleDeleteSemester(currentSemesterId)
-              setIsConfirmSemesterDeleteModalOpen(false)
-            }}
-            onCancel={() => setIsConfirmSemesterDeleteModalOpen(false)}
-          />
-        )}
-      </Card>
+      <SemesterManagement
+        orgId={organization?.id ?? -1}
+        organizationSemesters={organizationSemesters}
+        setOrganizationSemesters={setOrganizationSemesters}
+      />
     </div>
   ) : (
     <Spin />
